$rootScope.iveBeenFramed tells us if embedded
instead of basing the new window off of “source” which is added to the
session, we use the frame busting method to check if top != self and
set $rootScope.iveBeenFramed = true in app.js. This allows us to embed
the site anywhere, and if you click checkout it will open in a new
window.