[fix, OS, checkout join] jafrabiz Free shipping promo (pre-order), fixed VISA image alignment.
Nothing too fancy with the visuals of the modal. Please note that the &debug=true kept triggering the modal for start and finish only; perhaps it is a bug. But the modal appears upon the loading of Start step. Also I was unable to test for Finish/review step but feel free to refactor the code as needed.